# docs/complete-api-swagger.yaml
openapi: 3.0.0
info:
  title: 🛒 Ghanshyam Murti Bhandar - Complete Ecommerce API
  description: |
    # 🚀 Enterprise-Grade Ecommerce Platform API
    
    **239+ API Endpoints** covering all aspects of modern ecommerce for Android integration.
    
    ## 🔥 Complete Feature Set
    - **🔐 Authentication** - JWT-based secure authentication (4 endpoints)
    - **📦 Products** - Complete product management (12 endpoints)
    - **🏷️ Categories** - Category management (8 endpoints)
    - **🛒 Shopping Cart** - Real-time cart management (6 endpoints)
    - **❤️ Wishlist** - Save favorite products (5 endpoints)
    - **📍 Addresses** - Multi-address support (6 endpoints)
    - **📋 Orders** - Complete order lifecycle (15 endpoints)
    - **💳 Payments** - Razorpay, COD, Wallet (10 endpoints)
    - **🚚 Shipping** - Shiprocket integration (12 endpoints)
    - **💰 Wallet** - Digital wallet system (8 endpoints)
    - **🎫 Coupons** - Discount system (10 endpoints)
    - **⭐ Reviews** - Rating system (8 endpoints)
    - **👥 Users** - User management (10 endpoints)
    - **🏢 Admin** - Admin dashboard (25 endpoints)
    - **📊 Analytics** - Business intelligence (15 endpoints)
    - **📁 File Upload** - Image management (5 endpoints)
    - **🔔 Notifications** - Real-time updates (6 endpoints)
    - **🧾 Invoices** - GST compliance (8 endpoints)
    - **📦 Inventory** - Stock management (12 endpoints)
    - **🏪 Suppliers** - Vendor management (10 endpoints)
    
    ## 🎯 Android Integration Ready
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token for authentication. Get token from `/auth/login`.
        
        **Usage:** `Authorization: Bearer <your-jwt-token>`
        
        **Example:** `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Request success status
          example: true
        message:
          type: string
          description: Response message
          example: "Operation completed successfully"
        data:
          type: object
          description: Response data
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
          example: "2025-07-25T14:30:00Z"
      required:
        - success
        - message

    # Error Response
...
    Product:
      type: object
      properties:
        _id:
          type: string
          description: Product ID
          example: "60f1b2b3c4d5e6f7g8h9i0j2"
        name:
          type: string
          description: Product name
          example: "Ganesh Statue - Marble Finish"
        description:
          type: string
          description: Product description
          example: "Beautiful handcrafted Ganesh statue made from premium marble"
        price:
          type: number
          description: Product price in INR
          example: 2500
        originalPrice:
          type: number
          description: Original price before discount
          example: 3000
        discount:
          type: number
          description: Discount percentage
          example: 16.67
        category:
          $ref: '#/components/schemas/Category'
        subcategory:
          type: object
          properties:
            _id:
              type: string
            name:
              type: string
        stock:
          type: integer
          description: Available stock
          example: 10
        images:
          type: array
          items:
            type: string
          description: Product image URLs
          example: ["https://cdn.example.com/product1.jpg", "https://cdn.example.com/product2.jpg"]
        specifications:
          type: object
          description: Product specifications
          example:
            material: "Premium Marble"
            height: "12 inches"
            weight: "2 kg"
            color: "White"
        tags:
          type: array
          items:
            type: string
          description: Product tags
          example: ["ganesh", "statue", "marble", "religious"]
        rating:
          type: object
          properties:
            average:
              type: number
              example: 4.5
            count:
              type: integer
              example: 25
        isActive:
          type: boolean
          description: Product status
          example: true
        isFeatured:
          type: boolean
          description: Featured product status
          example: false
        seo:
          type: object
          properties:
            slug:
              type: string
              example: "ganesh-statue-marble-handcrafted"
            metaTitle:
              type: string
              example: "Premium Ganesh Statue - Handcrafted Marble"
            metaDescription:
              type: string
              example: "Buy premium handcrafted Ganesh statue made from marble"
        createdAt:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-07-25T14:30:00Z"

    # Category Schema
    Category:
      type: object
      properties:
        _id:
          type: string
          description: Category ID
          example: "60f1b2b3c4d5e6f7g8h9i0j3"
        name:
          type: string
          description: Category name
          example: "Statues"
        description:
          type: string
          description: Category description
          example: "Religious statues and idols"
        image:
          type: string
          description: Category image URL
          example: "https://cdn.example.com/categories/statues.jpg"
        isActive:
          type: boolean
          description: Category status
          example: true
        parent:
          type: string
          description: Parent category ID
          example: "60f1b2b3c4d5e6f7g8h9i0j4"
        children:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        productCount:
          type: integer
          description: Number of products in category
          example: 25
        seo:
          type: object
          properties:
            slug:
              type: string
              example: "religious-statues"
            metaTitle:
              type: string
              example: "Religious Statues - Handcrafted Idols"
        createdAt:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00Z"

    # Cart Schema
    Cart:
      type: object
      properties:
        _id:
          type: string
          description: Cart ID
          example: "60f1b2b3c4d5e6f7g8h9i0j5"
        user:
          type: string
          description: User ID
          example: "60f1b2b3c4d5e6f7g8h9i0j1"
        items:
          type: array
          items:
            type: object
            properties:
              product:
                $ref: '#/components/schemas/Product'
              quantity:
                type: integer
                example: 2
              price:
                type: number
                example: 2500
              total:
                type: number
                example: 5000
        pricing:
          type: object
          properties:
            subtotal:
              type: number
              example: 5000
            tax:
              type: number
              example: 900
            shipping:
              type: number
              example: 100
            discount:
              type: number
              example: 200
            total:
              type: number
              example: 5800
        updatedAt:
          type: string
          format: date-time
          example: "2025-07-25T14:30:00Z"
paths:
...
  /cart/add:
    post:
      tags:
        - 🛒 Shopping Cart
      summary: ➕ Add Item to Cart
      description: |
        Add a product to the shopping cart or update quantity if already exists.

        **Android Integration:**
        - Use from product detail screen
        - Show success animation
        - Update cart badge count
        - Handle stock validation
      operationId: addToCart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - productId
                - quantity
              properties:
                productId:
                  type: string
                  description: Product ID to add
                  example: "60f1b2b3c4d5e6f7g8h9i0j2"
                quantity:
                  type: integer
                  minimum: 1
                  maximum: 10
                  example: 2
                  description: Quantity to add
                variant:
                  type: string
                  description: Product variant (if applicable)
                  example: "medium"
            examples:
              add_product:
                summary: Add Product to Cart
                value:
                  productId: "60f1b2b3c4d5e6f7g8h9i0j2"
                  quantity: 2
      responses:
        '200':
          description: ✅ Item added to cart successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          cart:
                            $ref: '#/components/schemas/Cart'
                          addedItem:
                            type: object
                            properties:
                              product:
                                $ref: '#/components/schemas/Product'
                              quantity:
                                type: integer
                              total:
                                type: number
        '400':
          description: ❌ Invalid request or insufficient stock
        '401':
          description: ❌ Unauthorized
        '404':
          description: ❌ Product not found
        '500':
          description: ❌ Internal server error
  /cart/{productId}:
    put:
      tags:
        - 🛒 Shopping Cart
      summary: ✏️ Update Cart Item
      description: |
        Update quantity of a specific item in the cart.

        **Android Integration:**
        - Use from cart screen
        - Real-time quantity updates
        - Show updated pricing
        - Handle stock validation
      operationId: updateCartItem
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
          description: Product ID to update
          example: "60f1b2b3c4d5e6f7g8h9i0j2"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quantity
              properties:
                quantity:
                  type: integer
                  minimum: 1
                  maximum: 10
                  example: 3
                  description: New quantity
      responses:
        '200':
          description: ✅ Cart item updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          cart:
                            $ref: '#/components/schemas/Cart'
        '400':
          description: ❌ Invalid quantity or insufficient stock
        '401':
          description: ❌ Unauthorized
        '404':
          description: ❌ Product not found in cart
        '500':
          description: ❌ Internal server error

    delete:
      tags:
        - 🛒 Shopping Cart
      summary: 🗑️ Remove Item from Cart
      description: |
        Remove a specific item from the shopping cart.

        **Android Integration:**
        - Use from cart screen
        - Show confirmation dialog
        - Update cart totals
        - Handle empty cart state
      operationId: removeFromCart
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
          description: Product ID to remove
          example: "60f1b2b3c4d5e6f7g8h9i0j2"
      responses:
        '200':
          description: ✅ Item removed from cart successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          cart:
                            $ref: '#/components/schemas/Cart'
        '401':
          description: ❌ Unauthorized
        '404':
          description: ❌ Product not found in cart
        '500':
          description: ❌ Internal server error
  /cart/clear:
    delete:
      tags:
        - 🛒 Shopping Cart
      summary: 🧹 Clear Cart
      description: |
        Remove all items from the shopping cart.

        **Android Integration:**
        - Use for clear cart functionality
        - Show confirmation dialog
        - Handle empty cart state
      operationId: clearCart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ✅ Cart cleared successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: ❌ Unauthorized
        '500':
          description: ❌ Internal server error

  # 📋 ORDER MANAGEMENT ENDPOINTS
...